+++ tesseract-admin/src/main/java/admin/service/impl/TesseractLogServiceImpl.java
@@ -1,12 +1,15 @@
 package admin.service.impl;
 
+import admin.core.component.SendMailComponent;
 import admin.core.event.RetryEvent;
 import admin.entity.TesseractFiredJob;
 import admin.entity.TesseractLog;
+import admin.entity.TesseractTrigger;
 import admin.mapper.TesseractLogMapper;
 import admin.pojo.StatisticsLogDO;
 import admin.service.ITesseractFiredJobService;
 import admin.service.ITesseractLogService;
+import admin.service.ITesseractTriggerService;
 import admin.util.AdminUtils;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
@@ -25,6 +28,7 @@ import org.springframework.util.StringUtils;
 import tesseract.core.dto.TesseractAdminJobNotify;
 import tesseract.exception.TesseractException;
 
+import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.time.ZoneOffset;
 import java.time.temporal.ChronoUnit;
@@ -47,13 +51,16 @@ import static admin.constant.AdminConstant.LOG_SUCCESS;
 public class TesseractLogServiceImpl extends ServiceImpl<TesseractLogMapper, TesseractLog> implements ITesseractLogService {
 
     @Autowired
-    private ApplicationContext applicationContext;
-
-    @Autowired
     private ITesseractFiredJobService firedJobService;
 
     @Autowired
+    private SendMailComponent sendMailComponent;
+
+    @Autowired
     private EventBus retryEventBus;
+
+    @Autowired
+    private ITesseractTriggerService tesseractTriggerService;
 
     private int statisticsDays = 7;
 
@@ -62,26 +69,36 @@ public class TesseractLogServiceImpl extends ServiceImpl<TesseractLogMapper, Tes
     public void notify(TesseractAdminJobNotify tesseractAdminJobNotify) {
         Long logId = tesseractAdminJobNotify.getLogId();
         String exception = tesseractAdminJobNotify.getException();
-        Integer jobId = tesseractAdminJobNotify.getJobId();
+//        Integer jobId = tesseractAdminJobNotify.getJobId();
         TesseractLog tesseractLog = this.getById(logId);
+        @NotNull Integer triggerId = tesseractAdminJobNotify.getTriggerId();
+        TesseractTrigger tesseractTrigger = tesseractTriggerService.getById(triggerId);
         if (tesseractLog == null) {
             log.error("获取日志为空:{}", tesseractAdminJobNotify);
             throw new TesseractException("获取日志为空" + tesseractAdminJobNotify);
         }
+        QueryWrapper<TesseractFiredJob> firedJobQueryWrapper = new QueryWrapper<>();
+        firedJobQueryWrapper.lambda().eq(TesseractFiredJob::getTriggerId, tesseractAdminJobNotify.getTriggerId());
         if (!StringUtils.isEmpty(exception)) {
+            sendMailComponent.missionFailedSendMail(tesseractAdminJobNotify);
             tesseractLog.setStatus(LOG_FAIL);
             tesseractLog.setMsg(exception);
+            firedJobQueryWrapper.lambda().eq(TesseractFiredJob::getLogId, tesseractAdminJobNotify.getLogId());
+            TesseractFiredJob tesseractFiredJob = firedJobService.getOne(firedJobQueryWrapper);
+            if (tesseractTrigger.getRetryCount() > tesseractFiredJob.getRetryCount()) {
+                tesseractFiredJob.setRetryCount(tesseractFiredJob.getRetryCount() + 1);
+                firedJobService.updateById(tesseractFiredJob);
+            } else {
+                firedJobService.remove(firedJobQueryWrapper);
+            }
             //执行失败重试
             retry(tesseractAdminJobNotify);
         } else {
             tesseractLog.setStatus(LOG_SUCCESS);
             tesseractLog.setMsg("执行成功");
+            firedJobService.remove(firedJobQueryWrapper);
         }
         tesseractLog.setEndTime(System.currentTimeMillis());
-        //移除fired job
-        QueryWrapper<TesseractFiredJob> firedJobQueryWrapper = new QueryWrapper<>();
-        firedJobQueryWrapper.lambda().eq(TesseractFiredJob::getJobId, jobId);
-        firedJobService.remove(firedJobQueryWrapper);
         //更新日志状态
         this.updateById(tesseractLog);
     }
